Remediate dual-model review findings
Two independent adversarial reviews (Claude Opus 5, GPT-5.6-Sol) read M1 and M2
at f0b8c00. This is the correctness half of what they found. Everything here is
authored from the clause named in the change, and every fix is accompanied by a
mutation that puts the defect back and shows a named test going red.

Contract defects, where the code disagreed with Appendix B or Section 22:

- a quoted-namespace NUL value is SHELL001, not SERIALIZE001;
- an uncontainable or link-reached destination is PATH001, not PATH002 - nothing
  was opened, so it is not an open failure, and the difference is whether a
  reader retries or rewrites the scheme;
- PARSE001 is once per failing source, not once per malformed record;
- WILDCARD001 and REFERENCE001 are once per rule and once per owning value, so
  the column no longer splits one rule into several occurrences.

Ordering and shape defects:

- Section 5.2's "a wildcard token sorts after any literal text at the same
  position" is a position in the component's text, not in its token list.
  Comparing token against token made a* sort before ab, because the literals "a"
  and "ab" decided it and the wildcard was never reached. The comparison now
  walks scalars. A total-order property test cannot see this: the wrong order is
  still total.
- Section 4.4's "any later deep descendant refreshes the mapping shape-mark of
  every ancestor" was refreshing with the child's position mark, which Section
  5.2 pins to whatever first materialised the child. A three-level case in one
  source therefore lost the refresh entirely. NodeMarks.Latest carries it.
- a content-token address is a Section 5.4 ordering value and so is 64-bit;
  #2147483648 was rejected as too large.

Scheme compilation:

- output=json,json is a blocking scheme error under Section 17.5, not two
  contributions to one destination that then collide with themselves;
- two merge directives at one path are an override under Section 15.2. They were
  both retained and reached ImmutableDictionary.CreateRange as a duplicate key,
  which is a user-caused condition escaping as an unhandled exception - Section
  6.3 forbids exactly that. Both the compiler and the strategy map are now
  last-wins.

Publication now validates before it creates, rather than reporting an escape it
has already committed.

Specification amendment: Section 14.1 said an empty flat output is "exactly one
LF under Section 24" while Section 24 says a text output with no content is zero
bytes and "a single LF is never emitted merely to terminate nothing". Section
14.1 cited an authority that contradicted it, so Section 14.1 is corrected. The
implementation already emitted zero bytes; the point is that it could not have
been derived from the contract as written. Derived artifacts regenerated;
bundle r22 -> r23.

KNOWN-LIMITS section 3 claimed the publication path is "TOCTOU-safe by
construction". That is true of the spike and not of the shipped sink, which
resolves, checks and then opens because .NET exposes no no-follow open. The
claim is replaced with what the sink actually guarantees.

The containment check had no test that could reach it - the ancestor link
refusal always fired first - so a mutation replacing its exception type
survived. It is now exercised directly against the sink.

Copilot-Session: 28714a7f-5c88-4a3c-99ea-b9c0107e1155

+++ src/Namespace2Xml.Core/Profiles/NameParts.cs
@@ -105,8 +105,13 @@ public sealed record AttributePart(XmlNameComponent Name) : NamePart;
 /// <summary>
 /// An Appendix A.2 <c>typed-content</c>: <c>#n</c>, the Section 11.4 content-token ordering value.
 /// </summary>
-/// <param name="Ordinal">The ordering value. Never negative, and never written with a leading zero.</param>
-public sealed record ContentPart(int Ordinal) : NamePart;
+/// <param name="Ordinal">
+/// The ordering value. Never negative, and never written with a leading zero. Section 11.4 assigns
+/// content-token addresses "using Section 5.4", which makes them stable signed 64-bit ordering
+/// values through 9,223,372,036,854,775,807; a 32-bit ordinal would reject a valid address a
+/// sequence high-water allocator can legitimately reach.
+/// </param>
+public sealed record ContentPart(long Ordinal) : NamePart;
 
 /// <summary>Structural comparison and canonical form for token sequences.</summary>
 /// <remarks>
